KO html, passport routes, server.js clean up and front end jquery bones (jquery is WIP)
