+++ server.js
@@ -9,57 +9,15 @@ var LocalStrategy = require("passport-local").Strategy;
 var app = express();
 var PORT = process.env.PORT || 3000;
 
-// Middleware
+// // Middleware
 app.use(session({ secret: "consign" }));
 app.use(express.urlencoded({ extended: false }));
 app.use(express.json());
-app.use(express.static("public"));
+app.use(express.static("public")); //all static files, look inside the public  folder
 
-// Passport
-app.use(passport.initialize());
-app.use(passport.session());
 
-passport.use(new LocalStrategy(
-  function(username, password, done){
-    // TODO: check auth against database
-    if(username === "jdoe@example.com" && password === "open-sesame") {
-      return done(null, username);
-    }
-    return done(null, false);
-  }
-));
 
-passport.serializeUser((user, done) => {
-  done(null, user);
-});
-
-passport.deserializeUser((user, done) => {
-  done(null, user);
-});
-
-app.post("/login", 
-  passport.authenticate("local", { failureRedirect: "/login.html" }),
-  function(req, res) {
-    res.redirect("/");
-  });
-
-function isUserAuthenticated(req, res, next) {
-  if (req.user) {
-    console.log("auth succeeded")
-    next();
-  } else {
-    console.log("login please")
-    res.send('You must login!');
-  }
-}
-
-// TODO: for testing purposes
-app.get("/secret", isUserAuthenticated, (req, res) => {
-  res.send("Page to test authentication.");
-  console.log("secret route accessed by " + req.user);
-});
-
-// Logout route
+//Logout route
 app.get("/logout", (req, res) => {
   req.logout(); 
   res.redirect("/login.html");
@@ -77,6 +35,7 @@ app.set("view engine", "handlebars");
 // Routes
 require("./routes/apiRoutes")(app);
 require("./routes/htmlRoutes")(app);
+require("./routes/loginHtmlRoutes")(app);
 
 var syncOptions = { force: false };
